LayoutTests/imported/w3c/web-platform-tests/css/selectors/invalidation/attribute-expected.txt


PASS [att] selector is effective
PASS [att=val] selector is effective
PASS [att~=val] selector is effective
PASS [att|=val] selector is effective
PASS #id selector is effective
PASS .class selector is effective
